+++ NEWS.md
@@ -1,3 +1,5 @@
+# repana (development version)
+
 # NEWS repana
 
 ## repana 2.1.0